make raw try catch block import consistent to other imports
Usually we import a "raw" object that is stored within `ClassFileImportRecord`. For try-catch-blocks we somehow store the imported data within `ClassFileImportRecord` as a domain builder already instead of converting it to a domain builder later on. We now consolidate this to apply the same behavior when importing try-catch-blocks, i.e. store them as "raw" object and convert them to domain object builder during the domain object creation phase.
